Add context to record action
Reducer functions of each type only trigger if context is matched.

This fixes a bug when the field config of edit and show is different.
Edit page would trigger record fetch, and change the sate of all other
pages. Then there may be missing field when displaying show page after
that, which makes the app crash.